Initial fix for "Clang 3.9 and missing member definitions for template classes" (Issue 300) This tested OK under CLang 3.9, 3.6, GCC 4.8 and GCC 4.9. Windows will probably ned some additional tuning

+++ pubkey.h
@@ -327,7 +327,7 @@ public:
 
 // ********************************************************
 
-typedef std::pair<const byte *, size_t> HashIdentifier;
+typedef std::pair<const byte *, unsigned int> HashIdentifier;
 
 //! \class PK_SignatureMessageEncodingMethod
 //! \brief Interface for message encoding method for public key signature schemes.
